feat(PUPIL-291): lift fetching of the transcript file out of the query
This was a fun one. It only became apparent after I pushed to a branch
and CI failed that the curriculum API is imported in the client side
and not just on the server like the name `node-lib` would suggest. The
`fs` module among others (which the gcloud lib depends on) was causing
the next build to blow up. So i've moved it up to the page level so that
it can be properly tree shaken.