fix: lesson page blowing up when a download existence check fails
If the download existence check fails, we should not block the page from rendering.
There appear to be two ways the `getDownloadResourcesExistence` will report a missing
resource. Either by returning an object in the form
`{ resources: ["worksheet-pdf", { exists: false }] }`
or by throwing. Catching and reporting the error matches the behaviour of
the teachers downloads page and allows the lesson page to render