chore: use download canonical query for the worksheet download check
until now we were using the `getDownloadResourcesExistence` function
which hits the legacy or the new download API depending on whether the
lesson is a legacy or new lesson. The legacy endpoint will 404 when the
worksheet does not exist instead of returning a successful response with
the `exists` field set to `false` for the worksheet as expected. This
causes errors to be thrown and caught by Bugsnag which is not desirable.

We have a query `lessonDownloadsCanonical` which appears to be able to
fulfill the same function as `getDownloadResourcesExistence` without the
the munging of the response.