chore(LESQ-972): stop using edge middleware and use separate routes
I had trouble getting the edge middleware to run owing to the way that
we manage env vars. I've decided to leave the yak unshaven and use vanilla
api routes instead. If we go with Auth0 then we can deal with making things
run in edge functions.

i've also switched to creating new routes for the downloads pages that
require auth — this will make it easier to feature flag.