chore: replace middleware auth check with client side auth
Clerk does not support middleware auth protection on pages that use SSG.
So we have to stick with client-side authentication. This only stops rendering
of the page — page data is still available by hitting the SSG data directly from
Next's cache. In the case of the downloads pages this data is not priveleged.