fix: set the sign-in and sign-up routes on ClerkProvider
the production Clerk environment is misconfigured and both the sign-in
and sign-up routes point to labs.thenational.academy. It makes sense to
have these paths under version control so i'm setting them in code
instead of doing it through the Clerk UI to avoid causing any issues with
the AI project. Clerk themselves say you should do this and it keeps OWA
decoupled from other projects.