fix: role selection form breaking after adding `noValidate` to form
nnot sure what happenend here but the tests started failing on the
role selection form once I set `noValidate` on the form. I needed this
to disable the built in HTML validation that was triggered after I added
`required` to the school picker input. I suspect there was a race between
the validation triggered by the `onSubmit` handler (which is actually called
`onClick` on the "Continue" button) and the validation routine triggered
by `react-hook-form`. Essentially in the tests the invalid state created
by `onSubmit` was immediately overriden when the Zod validation was fired.

My solution was to pull all the validation into the Zod schema using `refine`
which seems to work well.