feat(LESQ-981): query for `geo_restricted` and `login_required` in downloads
this only updates the existing queries to get `geo_restricted` and
`login_required` for lessons and log it to the console when the feature flag is
enabled. This also removes the hard coded geo restriction enabler from the feature
flag as we need to do it anyway and it was easier than re-implementing.