fix: rendering the registration UI twice
rendering two sign-up/sign-in forms was causing the email confirmation
to be sent twice. I believe this was intended to make it easier to
present an alternative mobile view of the sign-up UI.

it appears to ultimately be an implementation bug in Clerk, because this
only happens with the confirmation link flow and not with the confirmation
code flow. Also if you refresh the page it will send another link which
is not great.